BOL-25 refactor: BaseRepository 에서 error 시 throw 대신 ErrorItem 을 던지도록 수정
