+++ domain/src/main/java/com/yapp/bol/domain/handle/BaseRepository.kt
@@ -1,6 +1,8 @@
-package com.yapp.bol.data.utils
+package com.yapp.bol.domain.handle
 
+import com.google.gson.Gson
 import com.yapp.bol.domain.model.ApiResult
+import com.yapp.bol.domain.model.ErrorItem
 import retrofit2.Response
 import java.io.IOException
 
@@ -13,18 +15,22 @@ abstract class BaseRepository {
         } catch (e: Exception) {
             return when (e) {
                 is IOException -> {
-                    ApiResult.Error(IOException(e.message ?: "Internet error runs"))
+                    throw IOException("Network Error")
                 }
 
-                else -> ApiResult.Error(e)
+                else -> throw Exception(e.message)
             }
         }
 
         val body = response.body()
+
         return if (body != null) {
             ApiResult.Success(body)
         } else {
-            ApiResult.Error(IOException("Body Null Error"))
+            val json = response.errorBody()?.string() ?: ""
+            ApiResult.Error(
+                Gson().fromJson(json, ErrorItem::class.java),
+            )
         }
     }
 }
