+++ domain/src/main/java/com/yapp/bol/domain/model/ErrorItem.kt
@@ -1,0 +1,6 @@
+package com.yapp.bol.domain.model
+
+data class ErrorItem(
+    val code: String,
+    val message: String,
+)
