+++ presentation/src/main/java/com/yapp/bol/presentation/utils/CheckedApiResult.kt
@@ -1,14 +1,15 @@
 package com.yapp.bol.presentation.utils
 
 import com.yapp.bol.domain.model.ApiResult
+import com.yapp.bol.domain.model.ErrorItem
 
 fun <T : Any> checkedApiResult(
     apiResult: ApiResult<T>,
     success: (T) -> Unit,
-    error: (Throwable) -> Unit,
+    error: ((ErrorItem) -> Unit)? = null,
 ) {
     when (apiResult) {
         is ApiResult.Success -> success(apiResult.data)
-        is ApiResult.Error -> error(apiResult.exception)
+        is ApiResult.Error -> error?.invoke(apiResult.exception)
     }
 }
